Updating settings window to have tabs
